Dynamically create portfolio items from markdown Prior to this change, there weren't links to any portfolio items.
This change allows markdown files to be queried for content. All
portfolio items will be sourced from markdown files. Markdown files
will contain a blurb about the item, a relative path to its image,
and if it's to be used on the homepage. Gatsby image is used to
dynamically create optimized images for different screen sizes.